+++ tests/test_modules/test_select_cohort_advanced/test_extract.py
@@ -9,7 +9,9 @@ from corebehrt.constants.data import (
     TIMESTAMP_COL,
     VALUE_COL,
 )
-from corebehrt.ozempic.criteria.extract import extract_patient_criteria
+from corebehrt.modules.cohort_handling.advanced.criteria.extract import (
+    extract_patient_criteria,
+)
 
 
 class TestCriteriaExtraction(unittest.TestCase):
